Move AnnotatedPyccelSymbol from internals to variable to avoid circular imports
