+++ pyccel/ast/core.py
@@ -22,7 +22,6 @@ from .datatypes import (datatype, DataType, NativeSymbol,
                         NativeBool, NativeRange,
                         NativeTuple, str_dtype)
 from .internals import Slice, PyccelSymbol, PyccelInternalFunction, get_final_precision
-from .internals import AnnotatedPyccelSymbol
 
 from .literals  import LiteralInteger, Nil, LiteralFalse
 from .literals  import NilArgument, LiteralTrue
@@ -31,7 +30,7 @@ from .operators import PyccelAdd, PyccelMinus, PyccelMul, PyccelDiv, PyccelMod, 
 from .operators import PyccelOperator, PyccelAssociativeParenthesis, PyccelIs
 
 from .variable import DottedName, IndexedElement
-from .variable import Variable
+from .variable import Variable, AnnotatedPyccelSymbol
 
 errors = Errors()
 pyccel_stage = PyccelStage()
