+++ pyccel/parser/scope.py
@@ -9,8 +9,8 @@
 from pyccel.ast.core      import ClassDef
 from pyccel.ast.headers   import MacroFunction, MacroVariable
 from pyccel.ast.headers   import FunctionHeader, MethodHeader
-from pyccel.ast.internals import PyccelSymbol, AnnotatedPyccelSymbol
-from pyccel.ast.variable  import Variable, DottedName
+from pyccel.ast.internals import PyccelSymbol
+from pyccel.ast.variable  import Variable, DottedName, AnnotatedPyccelSymbol
 
 from pyccel.parser.syntax.headers import FunctionHeaderStmt
